added sensors and gyro code to teleop
Added the DeviceInterfaceModule and the BNO055IMU gyro to the tele-op
code. Created initializeSensors() method that maps and initializes the
gyro. Added the InterruptedException tag to the runOpMode() method.
Created the getGyroMethod() which keeps the gyro going instead of
resetting to negative value at 180 degrees. Created the
composeTelemetry() method so that telemetry.update() could be called to
get the info instead of having to add the data everytime.